+++ source/SURF/R/simulationFunctions.R
@@ -1,7 +1,7 @@
 ### Title:    Data Simulation Functions for the SURF Package
 ### Author:   Kyle M. Lang
 ### Created:  2017-OCT-25
-### Modified: 2018-JUN-01
+### Modified: 2018-OCT-15
 
 ##-------------------- COPYRIGHT & LICENSING INFORMATION ---------------------##
 ##  Copyright (C) 2018 Kyle M. Lang <k.m.lang@uvt.nl>                         ##
@@ -22,36 +22,52 @@
 ##  with this program.  If not, see <http://www.gnu.org/licenses/>.           ##
 ##----------------------------------------------------------------------------##
 
+
+cleanInputs <- function() {
+    with(parent.frame(),
+    {
+        ## Check/modify the mean vector:
+        if(length(means) == 1) means <- rep(means, nVars)
+        else if(length(means) != nVars)
+            stop("'means' must have length 1 or length 'nVars'")
+
+        ## Check/modify the scales vector:
+        if(length(scales) == 1) scales <- rep(scales, nVars)
+        else if(length(scales) != nVars)
+            stop("'scales' must have length 1 or length 'nVars'")
+
+        ## Check/modify the covariance matrix:
+        if(is.vector(sigma) & length(sigma) == 1) {
+            ## Generate a covariance matrix from 'scales' and 'sigma':
+            w1 <- matrix(scales, nVars, nVars)
+            w2 <- matrix(scales, nVars, nVars, byrow = TRUE)
+            
+            maxCov <- w1 * w2
+            
+            sigma       <- maxCov * sigma
+            diag(sigma) <- scales^2
+        }
+        else if(!is.matrix(sigma))
+            stop("'sigma' must be a matrix or a length-one vector")
+    })
+}
+
+
+
 simRegData <- function(nObs,
-                       nPreds,
                        r2,
                        sigma,
                        beta,
+                       nVars           = ncol(sigma),
                        means           = 0,
                        scales          = 1,
                        itemsPerPred    = 1,
                        predReliability = 0.8)
 {
-    if(length(means) == 1) means <- rep(means, nPreds)
-
-    
-    if(is.matrix(sigma))
-        sigmaX <- sigma
-    else if(is.vector(sigma) & length(sigma) == 1) {
-        ## Generate a covariance matrix from 'scales' and 'collin':
-        w1 <- matrix(scales, nPreds, nPreds)
-        w2 <- matrix(scales, nPreds, nPreds, byrow = TRUE)
-        
-        maxCov <- w1 * w2
-        
-        sigmaX       <- maxCov * sigma
-        diag(sigmaX) <- scales^2
-    }
-    else
-        stop("'sigma' must be a matrix or a length-one vector")
+    cleanInputs()
     
     ## Simulate predictor data:
-    X <- cbind(1, rmvnorm(nObs, means, sigmaX))
+    X <- cbind(1, rmvnorm(nObs, means, sigma))
 
     ## Simulate the outcome:
     eta    <- X %*% beta
@@ -61,11 +77,11 @@ simRegData <- function(nObs,
     ## Generate a latent structure wherein each predictor is indicated by
     ## 'itemsPerPred' observed variables:
     if(itemsPerPred > 1) {
-        nItems   <- nPreds * itemsPerPred
-        loadings <- matrix(0, nItems, nPreds)
+        nItems   <- nVars * itemsPerPred
+        loadings <- matrix(0, nItems, nVars)
 
         ## Populate loading matrix:
-        for(m in 1 : nPreds) {
+        for(m in 1 : nVars) {
             for(n in 1 : itemsPerPred) {
                 offset                  <- (m - 1) * itemsPerPred
                 loadings[n + offset, m] <- sqrt(predReliability)
@@ -79,20 +95,33 @@ simRegData <- function(nObs,
         X <- X[ , -1] %*% t(loadings) + rmvnorm(nObs, rep(0, nItems), theta)
         
         outDat           <- data.frame(y, X)
-        colnames(outDat) <- c("y", paste0("x", c(1 : nItems)))
+        colnames(outDat) <- c("y", paste0("x", 1 : nItems))
     }
     else {
         outDat           <- data.frame(y, X[ , -1])
-        colnames(outDat) <- c("y", paste0("x", c(1 : nPreds)))
+        colnames(outDat) <- c("y", paste0("x", 1 : nVars))
     }
     outDat
-}# END simulateData()
+}# END simRegData()
+
+
+
+simCovData <- function(nObs, sigma, nVars = ncol(sigma), means = 0, scales = 1)
+{
+    cleanInputs()
+    
+    ## Simulate the data:
+    outData           <- as.data.frame(rmvnorm(nObs, means, sigma))
+    colnames(outData) <- paste0("x", 1 : nVars)
+
+    outData
+}# END simCovData()
 
 
 
 makeRVec <- function(linPred, pm, snr, pattern) {
     ## Generate a linear predictor of missingness:
-    noise   <- sd(linPred) * (1 / snr)
+    noise   <- sd(linPred) / snr
     linPred <- linPred + rnorm(length(linPred), 0, noise)
 
     ## Use a probit model to simulate response propensities:
@@ -114,11 +143,22 @@ makeRVec <- function(linPred, pm, snr, pattern) {
     list(rVec = rVec, pattern = pattern)
 }# END makeRVec()
 
-        
+
+
+## Check for various kinds of trival list entries:
+empty <- function(x)
+    missing(x) || is.null(x) || is.na(x) || length(x) == 0 || x == ""
+
+
 
 imposeMissData <- function(data, targets, preds, pm, snr, pattern = "random") {
+    ## Exclude empty 'targets' slots:
+    targets <- targets[!sapply(targets, empty)]
+    
     ## Which mechanisms should be simulated?
-    mechFlag <- !is.na(targets)
+    mechs           <- c("mar", "mcar", "mnar")
+    mechFlag        <- mechs %in% names(targets)
+    names(mechFlag) <- mechs
 
     if(length(pattern) == 1) {
         pattern        <- rep(pattern, length(c(targets$mar, targets$mnar)))