+++ CHANGELOG.md
@@ -4,7 +4,17 @@ All notable changes to the **SURF** project will be documented in this file.
 The format is based on [Keep a Changelog][kacl], and this project adheres to
 [Semantic Versioning][sv].
 
-## 0.0.0.9003 - 2018-06-01 - ACTIVE
+## 0.0.0.9004 - 2018-10-15 - ACTIVE
+
+### Added
+- A new function, `simCovData`, to simulate data according to a basic covariance 
+  structure.
+
+### Changed
+- Improved implementation of `simRegData` and `imposeMissData` functions.
+- The `nPreds` argument in `simRegData` is now `nVars`.
+
+## 0.0.0.9003 - 2018-06-01
 
 ### Changed
 - Changed the `collin` argument to `sigma` in `simRegData`.